wget; keep old files
Added install of ca-certificates libustream-openssl in order to get wget working with ssl.
Added backup of old files

+++ root/install-nordvpn.sh
@@ -15,10 +15,10 @@ Ninstallopkg() { #opkgfilename #[temp]
   local T=$(opkg list-installed | cut -d ' ' -f 1 | grep "$1")
   if [ -z "$T" ]; then
     echo "I didn't find $1. Installing."
-    if [ "$2" = "temp" ]; then
+    if [ "$2" = "temp" ]; then 
       REMOVETEMP="$REMOVEOPKGTEMP $1"
-	else
-	  REMOVE="$REMOVEOPKGFAIL $1"
+    else
+      REMOVE="$REMOVEOPKGFAIL $1"
     fi
     opkg install $1
   fi
@@ -26,9 +26,7 @@ Ninstallopkg() { #opkgfilename #[temp]
 
 # It'll remove a list of opkg f=' '
 Nremoveopkg() { #listOfOpkg
-  if [ ! -z "$1" ]; then
-    opkg --autoremove remove $1
-  fi
+  [ ! -z "$1" ] && opkg --autoremove remove $1
 }
 
 # It'll remove a list of files f='\n'
@@ -37,19 +35,17 @@ Nremovefiles() { #listOfFiles #[bak]
   if [ ! -z "$1" ]; then
     for RF in $1; do
       if [ "$2" = "bak" ]; then
-	    mv $RF.backup $RF
-	  else
-	    rm $RF
-	  fi
+        mv $RF.backup $RF
+      else
+        rm $RF
+      fi
     done
   fi
 }
 
 Ncfile() {
 # file text
-  if [ ! -e "$1" ]; then
-    touch "$1"
-  fi
+  [ ! -e "$1" ] && touch "$1"
 ###?????????
 
 }
@@ -58,14 +54,17 @@ Ncfile() {
 Nfail() {
   local F=$1
   if [ ! $F ]; then
-    echo -e "Something went wrong. It'll abort.\nCleaning files."
-	Nremoveopkg $REMOVETEMP
-	Nremoveopkg $REMOVEFAIL
-	Nremovefiles $REMOVEBAK bak
-	Nremovefiles $REMOVEFILES
-	echo -e "Removing Nordvpn OVPN files. It'll take some time. Be patient"
-	rm -rf /etc/openvpn/nordvpn
-	exit $F
+    echo -e "Something went wrong. I'll abort.\nCleaning files."
+    Nremoveopkg $REMOVETEMP
+    Nremoveopkg $REMOVEFAIL
+    Nremovefiles $REMOVEBAK bak
+    Nremovefiles $REMOVEFILES
+    echo -e "Removing Nordvpn OVPN files. It'll take some time. Be patient"
+    if [ -d /etc/openvpn/nordvpn.old ]; then
+      [ -d /etc/openvpn/nordvpn ] && rm -rf /etc/openvpn/nordvpn
+      mv /etc/openvpn/nordvpn.old /etc/openvpn/nordvpn
+    fi
+    exit $F
   fi
 }
 
@@ -73,7 +72,7 @@ Nfail() {
 Ndownload() { #path #url-nofile #filename
   if [ -e $1/$3 ]; then
     mv $1/$3 $1/$3.backup
-	REMOVEBAK="$REMOVEBAK\n$1/$3"
+    REMOVEBAK="$REMOVEBAK\n$1/$3"
   else
     REMOVEFILES="$REMOVEFILES\n$1/$3"
   fi
@@ -86,10 +85,13 @@ Ndownload() { #path #url-nofile #filename
 echo "Updating opkg"
 opkg update
 Nfail $?
-Ninstallopkg unzip temp
-Ninstallopkg curl temp
 Ninstallopkg openvpn-openssl
 Ninstallopkg ip-full
+Ninstallopkg unzip temp
+Ninstallopkg curl temp
+Ninstallopkg wget temp
+Ninstallopkg ca-certificates temp
+Ninstallopkg libustream-openssl temp
 if [ -e /etc/init.d/uhttpd ]; then
   Ninstallopkg luci-app-openvpn
 fi
@@ -97,23 +99,25 @@ fi
 # Download Nordvpn Config files
 echo "If you have a slow connection, be patient."
 echo "Downloading official nordvpn files."
-cd /root ; curl -O --url "http://downloads.nordcdn.com/configs/archives/servers/ovpn.zip"
+cd /root
+curl -O --url "http://downloads.nordcdn.com/configs/archives/servers/ovpn.zip"
 ckfail $?
+if [ -d /etc/openvpn/nordvpn ]; then
+  [ -d /etc/openvpn/nordvpn.old ] && rm -rf /etc/openvpn/nordvpn.old
+  mv /etc/openvpn/nordvpn /etc/openvpn/nordvpn.old
+fi
 unzip -j -o /root/ovpn.zip -d /etc/openvpn/nordvpn/
 ckfail $?
 rm /root/ovpn.zip
 
-
-# Simplify name and content
+# Simplifying name and content
 for F in /etc/openvpn/nordvpn/*.ovpn ; do
-
 	#ls -1 /etc/openvpn/nordvpn/*.ovpn | sed -r 's:/etc/openvpn/nordvpn/(.*)\.nordvpn\.com\.(tcp|udp)\.ovpn:\1.\2:' -e 's:(.*)-tor([0-9])\.(tcp|udp):xTor\2-\1.\3:' -e 's:([a-z]{2}-[a-z]{2})([0-9])\.(tcp|udp):xDouble\2-\1.\3:' -e 's:_:-:g'
-
   NF=$( echo $F | sed -r \
-	-e 's:/etc/openvpn/nordvpn/(.+)\.nordvpn\.com\.(tcp|udp)\.ovpn:\1.\2:' \
-	-e 's:_:-:g' \
-	-e 's:(.+)-tor([0-9])\.(tcp|udp):xTor\2_\1.\3:' \
-	-e 's:([a-z]{2}-[a-z]{2})([0-9])\.(tcp|udp):xDouble\2_\1.\3:' )
+  -e 's:/etc/openvpn/nordvpn/(.+)\.nordvpn\.com\.(tcp|udp)\.ovpn:\1.\2:' \
+  -e 's:_:-:g' \
+  -e 's:(.+)-tor([0-9])\.(tcp|udp):xTor\2_\1.\3:' \
+  -e 's:([a-z]{2}-[a-z]{2})([0-9])\.(tcp|udp):xDouble\2_\1.\3:' )
 
   if [ -z "$NF" ]; then 
     continue; 
@@ -132,20 +136,20 @@ for F in /etc/openvpn/nordvpn/*.ovpn ; do
 done
 
 if [ -e "/etc/openvpn/nordvpn-auth" ]; then
-	CHOICE = ""
-	read -r -N 1 -t 30 -p "There is nordvpn credentials file already. Do you wish to change it? (y/n)" CHOICE
+  CHOICE = ""
+  read -r -N 1 -t 30 -p "There is nordvpn credentials file already. Do you wish to change it? (y/n)" CHOICE
 else
-	CHOICE = "y"
+  CHOICE = "y"
 fi
 
 if [ "$CHOICE" = "y" -o "$CHOICE" = "Y" ]; then
-	read -p 'What is your Nordvpn username (email)? ' uservar
-	read -sp 'What is your Nordvpn password? ' passvar
-	echo "I will write them to /etc/openvpn/nordvpn-auth"
-	echo "Edit it if you need"
-	echo $uservar > /etc/openvpn/nordvpn-auth
-	echo $passvar >> /etc/openvpn/nordvpn-auth
-	chmod u=rw,go= /etc/openvpn/nordvpn-auth
+  read -p 'What is your Nordvpn username (email)? ' uservar
+  read -sp 'What is your Nordvpn password? ' passvar
+  echo "I will write them to /etc/openvpn/nordvpn-auth"
+  echo "Edit it if you need"
+  echo $uservar > /etc/openvpn/nordvpn-auth
+  echo $passvar >> /etc/openvpn/nordvpn-auth
+  chmod u=rw,go= /etc/openvpn/nordvpn-auth
 fi
 
 cd /etc/openvpn
